refactor(User) : User의 cards 필드를 Hand라는 일급컬렉션으로 변경
